Need to get children folder in the loop

+++ api/hyperv-winrm/vhd.go
@@ -81,7 +81,6 @@ function Expand-Downloads {
     )
     process {
 		Push-Location $FolderPath
-		$vhdPath = Get-ChildItem $tempPath *"Virtual Hard Disks"* -Recurse -Directory
 
         get-item *.zip | % {
 			$tempPath = join-path $FolderPath "temp"
@@ -97,6 +96,8 @@ function Expand-Downloads {
     			}
             	[System.IO.Compression.ZipFile]::ExtractToDirectory($_.FullName, $tempPath)
 			}
+
+			$vhdPath = Get-ChildItem $tempPath *"Virtual Hard Disks"* -Recurse -Directory
 
             if (Test-Path $($vhdPath.FullName)) {
         		Move-Item "$($vhdPath.FullName)\*.*" $FolderPath
@@ -117,7 +118,9 @@ function Expand-Downloads {
 			$command = """$7zPath"" x ""$($_.FullName)"" -o""$tempPath""" 
 			& cmd.exe /C $command
 
-            if (Test-Path $($vhdPath.FullName)) {
+			$vhdPath = Get-ChildItem $tempPath *"Virtual Hard Disks"* -Recurse -Directory
+
+            if ($vhdPath -and Test-Path $($vhdPath.FullName)) {
         		Move-Item "$($vhdPath.FullName)\*.*" $FolderPath
 			} else {
 				Move-Item "$tempPath\*.*" $FolderPath
@@ -140,7 +143,9 @@ function Expand-Downloads {
 			$command = """$tarPath"" -C ""$tempPath"" -x -f ""$($_.FullName)"""
 			& cmd.exe /C $command
 
-            if (Test-Path $($vhdPath.FullName)) {
+			$vhdPath = Get-ChildItem $tempPath *"Virtual Hard Disks"* -Recurse -Directory
+
+            if ($vhdPath -and Test-Path $($vhdPath.FullName)) {
         		Move-Item "$($vhdPath.FullName)\*.*" $FolderPath
 			} else {
 				Move-Item "$tempPath\*.*" $FolderPath
